refactor(observations): colocate evaluator badge
